Refactor FakeRosterDataSource to be a class and inject it
The `FakeRosterDataSource` has been changed from an `object` to a `class`. This allows it to be injected as a dependency into ViewModels.

The following ViewModels have been updated to accept `OneRosterRosterDataSource` as a constructor parameter:
- `ClazzEditViewModel`
- `ClazzListViewModel`
- `ClazzDetailViewModel`

A Koin binding for `OneRosterRosterDataSource` has been added in `AppKoinModule.kt`, providing an instance of `FakeRosterDataSource`.